better cache performance

+++ GBDT/src/lib/util.cpp
@@ -14,7 +14,7 @@ using namespace std;
 
 namespace zyuco {
 	namespace Data {
-		LibSVMData fromLibSVM(const std::string & path, int featureCount) {
+		LibSVMData fromLibSVM(const std::string & path, size_t featureCount) {
 			DataFrame x;
 			DataColumn y;
 
@@ -77,7 +77,7 @@ namespace zyuco {
 			return { { move(trainX), move(trainY) }, { move(testX), move(testY) } };
 		}
 
-		std::pair<DataRow, double> parseLibSVMLine(const std::string & line, int featureCount) {
+		std::pair<DataRow, double> parseLibSVMLine(const std::string & line, size_t featureCount) {
 			vector<double> values(featureCount);
 			size_t label, index;
 			double value;
